requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain for text processing
langchain==0.1.0
langchain-community==0.0.10

# Text extraction
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0

# OpenAI
openai==1.3.7

# AWS S3
boto3==1.29.7

# Vector database
faiss-cpu==1.7.4
pymongo==4.6.0

# PDF generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Type hints
typing-extensions==4.8.0

